// embedded-components/src/api/generated/ep-transactions-v3.schemas.ts
export type ApiErrorReasonV2Location =
  | (typeof ApiErrorReasonV2Location)[keyof typeof ApiErrorReasonV2Location]
  | null;

export const ApiErrorReasonV2Location = {
  BODY: 'BODY',
  QUERY: 'QUERY',
  PATH: 'PATH',
  HEADER: 'HEADER',
} as const;

/**
 * Schema representing detailed reasons for an API error, including code, location, and message.
 */
export interface ApiErrorReasonV2 {
  /**
   * Short code that identifies the error - publicly cataloged and documented
   * @minLength 0
   * @maxLength 5
   * @nullable
   */
  code?: string | null;
  /**
   * Part of the request which is responsible for the reason
   * @nullable
   */
  location?: ApiErrorReasonV2Location;
  /**
   * The location of the property or parameter in error
   * @minLength 0
   * @maxLength 512
   * @nullable
   */
  field?: string | null;
  /**
   * Message describing the reason. This message can typically be displayed to your platform's users, except in cases specified otherwise
   * @minLength 0
   * @maxLength 2048
   */
  message: string;
}

/**
 * Schema representing a version 2 API error, including title, HTTP status, and additional context.
 */
export interface ApiErrorV2 {
  /**
   * Short humanly-readable title of the error
   * @minLength 1
   * @maxLength 512
   */
  title: string;
  /**
   * HTTP status code
   * @minimum 100
   * @maximum 599
   */
  httpStatus?: number;
  /**
   * Internal assigned traced identifier
   * @minLength 0
   * @maxLength 512
   */
  traceId?: string;
  /**
   * Client provided request identifier
   * @minLength 0
   * @maxLength 512
   */
  requestId?: string;
  /**
   * Provides additional context and detail on the validation errors
   * @minItems 0
   * @maxItems 100
   */
  context?: ApiErrorReasonV2[];
}